Consolidate util files into a folder
