+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './modules/repl'
+export * from './modules/repl-server'
